pyproject.toml
[tool.poetry]
name = "vitaerules"
version = "0.1.0"
description = "CrewAI-first personal assistant with memory, task management, and integrations"
authors = ["JuanLuBcn"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
crewai = "^0.86.0"
crewai-tools = "^0.17.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-telegram-bot = "^21.6"
chromadb = "^0.5.0"
openai = "^1.51.0"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
langchain-openai = "^0.2.2"
faster-whisper = "^1.0.3"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
ruff = "^0.6.9"
pre-commit = "^3.8.0"
ipython = "^8.28.0"

[tool.poetry.scripts]
vitaerules = "app.main:main"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=src/app --cov-report=term-missing --cov-report=html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
